research_copilot_webserver/requirements.txt
fastapi
uvicorn
requests
crewai
crewai-tools
pypdf
pydantic
google-generativeai
pathlib
pytest
pytest_cov
pytest-mock